fix de generacion de vuelo

+++ src/AerolineaFrba/Abm Aeronave/Aeronave.cs
@@ -11,7 +11,7 @@ namespace AerolineaFrba.Abm_Aeronave
         public int id { get; set; }
         public string matricula { get; set; }
         public string modelo { get; set; }
-        public int kgs_disponibles { get; set; }
+        public float kgs_disponibles { get; set; }
         public int butacas_totales { get; set; }
         public string fabricante { get; set; }
         public string servicio { get; set; }
